chore: refactor components to sfc
